requirements.txt
streamlit
python-docx
PyMuPDF
python-dotenv
google-generativeai
langdetect
docx2txt
spacy
